requirements-dev.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
google-api-core==1.14.3
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-cloud-vision==0.39.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
idna==2.8
image==1.5.27
olefile==0.46
Pillow==6.0.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.6
python-decouple==3.1
